feat: refactor db tests
