+++ packages/adapters/database/jest.config.js
@@ -2,8 +2,8 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   displayName: 'Database Adapter',
-  testMatch: ['<rootDir>/test/**/*.test.ts'],
-  globalSetup: '<rootDir>/test/jest.setup.ts',
+  testMatch: ['<rootDir>/test/**/*.spec.ts'],
+  globalSetup: '<rootDir>/test/setup.ts',
   collectCoverageFrom: ['src/**/*.ts', '!src/**/*.d.ts'],
   coverageDirectory: 'coverage',
   coverageReporters: ['text', 'lcov', 'html'],
