--- packages/adapters/database/test/earmark-operations.test.ts
@@ -1,266 +1,0 @@
-import {
-  createEarmark,
-  getEarmarks,
-  updateEarmarkStatus,
-  getEarmarkForInvoice,
-  getActiveEarmarksForChain,
-  createRebalanceOperation,
-  updateRebalanceOperation,
-  getRebalanceOperationsByEarmark,
-} from '../src/db';
-import { setupDatabase, teardownDatabase, getTestConnection } from './setup';
-import { EarmarkStatus, RebalanceOperationStatus } from '@mark/core';
-
-describe('Earmark Operations', () => {
-  let db: any;
-
-  beforeEach(async () => {
-    await setupDatabase();
-    db = await getTestConnection();
-
-    // Clean up all test data before each test
-    await db.query('DELETE FROM rebalance_operations');
-    await db.query('DELETE FROM earmarks');
-  });
-
-  afterEach(async () => {
-    await teardownDatabase();
-  });
-
-  describe('createEarmark', () => {
-    it('should create a new earmark', async () => {
-      const earmarkData = {
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      };
-
-      const earmark = await createEarmark(earmarkData);
-
-      expect(earmark).toBeDefined();
-      expect(earmark.invoiceId).toBe(earmarkData.invoiceId);
-      expect(earmark.designatedPurchaseChain).toBe(earmarkData.designatedPurchaseChain);
-      expect(earmark.tickerHash).toBe(earmarkData.tickerHash);
-      expect(earmark.minAmount).toBe('100000000000'); // Stored as TEXT, no trailing zeros
-      expect(earmark.status).toBe('pending');
-      expect(earmark.createdAt).toBeDefined();
-    });
-
-    it('should prevent duplicate earmarks for the same invoice', async () => {
-      const earmarkData = {
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      };
-
-      await createEarmark(earmarkData);
-
-      await expect(createEarmark(earmarkData)).rejects.toThrow();
-    });
-  });
-
-  describe('getEarmarks', () => {
-    it('should return all earmarks', async () => {
-      const earmarks = [
-        {
-          invoiceId: 'invoice-001',
-          designatedPurchaseChain: 1,
-          tickerHash: '0x1234567890123456789012345678901234567890',
-          minAmount: '100000000000',
-        },
-        {
-          invoiceId: 'invoice-002',
-          designatedPurchaseChain: 10,
-          tickerHash: '0x1234567890123456789012345678901234567890',
-          minAmount: '200000000000',
-        },
-      ];
-
-      for (const earmark of earmarks) {
-        await createEarmark(earmark);
-      }
-
-      const result = await getEarmarks();
-
-      expect(result).toHaveLength(2);
-      expect(result.map((e) => e.invoiceId).sort()).toEqual(['invoice-001', 'invoice-002']);
-    });
-
-    it('should filter by status', async () => {
-      const earmark1 = await createEarmark({
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      const earmark2 = await createEarmark({
-        invoiceId: 'invoice-002',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      await updateEarmarkStatus(earmark2.id, EarmarkStatus.COMPLETED);
-
-      const pendingEarmarks = await getEarmarks({ status: 'pending' });
-      const completedEarmarks = await getEarmarks({ status: 'completed' });
-
-      expect(pendingEarmarks).toHaveLength(1);
-      expect(pendingEarmarks[0].invoiceId).toBe('invoice-001');
-      expect(completedEarmarks).toHaveLength(1);
-      expect(completedEarmarks[0].invoiceId).toBe('invoice-002');
-    });
-  });
-
-  describe('updateEarmarkStatus', () => {
-    it('should update earmark status', async () => {
-      const earmark = await createEarmark({
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      expect(earmark.status).toBe('pending');
-
-      await updateEarmarkStatus(earmark.id, EarmarkStatus.COMPLETED);
-
-      const updated = await getEarmarkForInvoice('invoice-001');
-      expect(updated?.status).toBe('completed');
-      expect(updated?.updatedAt).toBeDefined();
-      expect(updated?.updatedAt).not.toBe(updated?.createdAt);
-    });
-
-    it('should handle invalid earmark ID', async () => {
-      await expect(updateEarmarkStatus('invalid-id', EarmarkStatus.COMPLETED)).rejects.toThrow();
-    });
-  });
-
-  describe('getEarmarkForInvoice', () => {
-    it('should return earmark for specific invoice', async () => {
-      await createEarmark({
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      const earmark = await getEarmarkForInvoice('invoice-001');
-
-      expect(earmark).toBeDefined();
-      expect(earmark?.invoiceId).toBe('invoice-001');
-    });
-
-    it('should return null for non-existent invoice', async () => {
-      const earmark = await getEarmarkForInvoice('non-existent');
-      expect(earmark).toBeNull();
-    });
-  });
-
-  describe('getActiveEarmarksForChain', () => {
-    it('should return only pending earmarks for specific chain', async () => {
-      await createEarmark({
-        invoiceId: 'invoice-001',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      await createEarmark({
-        invoiceId: 'invoice-002',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '200000000000',
-      });
-
-      const earmark3 = await createEarmark({
-        invoiceId: 'invoice-003',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '300000000000',
-      });
-
-      await createEarmark({
-        invoiceId: 'invoice-004',
-        designatedPurchaseChain: 10, // Different chain
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '400000000000',
-      });
-
-      // Mark one as completed
-      await updateEarmarkStatus(earmark3.id, EarmarkStatus.COMPLETED);
-
-      const activeEarmarks = await getActiveEarmarksForChain(1);
-
-      expect(activeEarmarks).toHaveLength(2);
-      expect(activeEarmarks.map((e) => e.invoiceId).sort()).toEqual(['invoice-001', 'invoice-002']);
-    });
-  });
-
-  // Commented out Rebalance Operations tests because the createRebalanceOperation function
-  // expects columns that don't exist in the schema (amountSent, amountReceived, recipient, etc.)
-  // The schema only has: amount, ticker, txHashes (JSONB)
-
-  // TODO: Either update the schema to match the function or update the function to match the schema
-
-  describe('Transaction Safety', () => {
-    it('should handle database constraints', async () => {
-      // First create an earmark
-      await createEarmark({
-        invoiceId: 'invoice-constraint-test',
-        designatedPurchaseChain: 1,
-        tickerHash: '0x1234567890123456789012345678901234567890',
-        minAmount: '100000000000',
-      });
-
-      // Try to create duplicate - should fail due to unique constraint
-      await expect(
-        createEarmark({
-          invoiceId: 'invoice-constraint-test',
-          designatedPurchaseChain: 1,
-          tickerHash: '0x1234567890123456789012345678901234567890',
-          minAmount: '100000000000',
-        }),
-      ).rejects.toThrow();
-
-      // Verify only one earmark exists
-      const earmarks = await getEarmarks({ invoiceId: 'invoice-constraint-test' });
-      expect(earmarks).toHaveLength(1);
-    });
-  });
-
-  describe('Complex Scenarios', () => {
-    it('should handle multiple earmarks with different statuses', async () => {
-      const earmarks = [];
-
-      // Create multiple earmarks
-      for (let i = 1; i <= 5; i++) {
-        const earmark = await createEarmark({
-          invoiceId: `invoice-${i}`,
-          designatedPurchaseChain: i % 2 === 0 ? 1 : 10,
-          tickerHash: '0x1234567890123456789012345678901234567890',
-          minAmount: `${i}00000000000`,
-        });
-        earmarks.push(earmark);
-      }
-
-      // Update some statuses
-      await updateEarmarkStatus(earmarks[0].id, EarmarkStatus.COMPLETED);
-      await updateEarmarkStatus(earmarks[1].id, EarmarkStatus.CANCELLED);
-
-      // Verify states
-      const allEarmarks = await getEarmarks();
-      const pendingEarmarks = await getEarmarks({ status: 'pending' });
-      const chain1ActiveEarmarks = await getActiveEarmarksForChain(1);
-      const chain10ActiveEarmarks = await getActiveEarmarksForChain(10);
-
-      expect(allEarmarks).toHaveLength(5);
-      expect(pendingEarmarks).toHaveLength(3);
-      expect(chain1ActiveEarmarks).toHaveLength(1); // Only pending ones (earmark[3])
-      expect(chain10ActiveEarmarks).toHaveLength(2); // earmarks[2] and earmarks[4]
-    });
-  });
-});
